feat: support profiling

+++ scripts/run-all.sh
@@ -12,16 +12,24 @@ for mul in $MUL_IMPLS; do
     for rsa in $RSA_IMPLS; do
       name="mul-${mul}_modmul-${modmul}_rsa-${rsa}"
       log="out/${name}.log"
+      prof="out/${name}.prof"
 
-      if [ -f "$log" ]; then
+      if [ -f "$log" ] && [ -f "$prof" ]; then
         echo "==> $name exists, skipping"
         continue
       fi
 
       echo "==> $name"
+
       make clean >/dev/null
       make MUL="$mul" MODMUL="$modmul" RSA="$rsa" >/dev/null
+
+      make clean >/dev/null
+      make MUL="$mul" MODMUL="$modmul" RSA="$rsa" 'RELCFLAGS=-O3 -DNDEBUG -pg' >/dev/null
+
       ./release/main >"$log" 2>&1
+      gprof ./release/main gmon.out >"$prof"
+      rm gmon.out
     done
   done
 done
